# tools/dashboard/backend/docs/swagger.yaml
definitions:
...
  xstores.MessageResponseDTO:
    description: Simple response with a message field.
    properties:
      message:
        description: Response message
        example: Operation completed successfully
        type: string
    type: object
  xstores.ObjectMetaDTO:
    description: Kubernetes resource metadata including name, namespace, labels and
      annotations.
    properties:
      annotations:
        additionalProperties:
          type: string
        description: Annotations attached to the resource
        type: object
      creationTimestamp:
        description: Creation timestamp
        type: string
      generation:
        description: Generation number, incremented on spec changes
        example: 1
        type: integer
      labels:
        additionalProperties:
          type: string
        description: Labels attached to the resource
        type: object
      name:
        description: Resource name, unique within namespace
        example: my-xstore
        type: string
      namespace:
        description: Kubernetes namespace
        example: default
        type: string
      resourceVersion:
        description: Resource version for optimistic concurrency
        example: "12345"
        type: string
      uid:
        description: Unique identifier for the resource
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
    type: object
  xstores.RebuildProgressDTO:
    description: Detailed progress of an XStore rebuild operation.
    properties:
      currentStep:
        description: Current step name
        example: CopyingData
        type: string
      eta:
        description: Estimated time remaining
        example: 10m
        type: string
      message:
        description: Status message
        example: Copying data from leader
        type: string
      percentage:
        description: Progress percentage (0-100)
        example: 60
        type: integer
      stepNumber:
        description: Step number
        example: 3
        type: integer
      targetPod:
        description: Target pod being rebuilt
        example: my-xstore-1
        type: string
      totalSteps:
        description: Total steps
        example: 5
        type: integer
    type: object
  xstores.RebuildStatusDTO:
    description: Status of an XStore rebuild operation.
    properties:
      endTime:
        description: End time of rebuild
        type: string
      error:
        description: Error message if failed
        type: string
      phase:
        description: Current rebuild phase
        example: Completed
        type: string
      progress:
        description: Progress percentage (0-100)
        example: 100
        type: integer
      rebuilding:
        description: Whether rebuild is in progress
        example: false
        type: boolean
      startTime:
        description: Start time of rebuild
        type: string
      targetPod:
        description: Pod being rebuilt
        example: my-xstore-1
        type: string
    type: object
  xstores.ResourceListDTO:
    description: Resource quantities for CPU and memory.
    properties:
      cpu:
        description: CPU resource (e.g., "2", "500m")
        example: "2"
        type: string
      memory:
        description: Memory resource (e.g., "4Gi", "512Mi")
        example: 4Gi
        type: string
    type: object
  xstores.ResourceRequirementsDTO:
    description: CPU and memory resource requirements.
    properties:
      limits:
        allOf:
        - $ref: '#/definitions/xstores.ResourceListDTO'
        description: Resource limits
      requests:
        allOf:
        - $ref: '#/definitions/xstores.ResourceListDTO'
        description: Resource requests
    type: object
...
  xstores.XStoreFollowerDTO:
    description: XStoreFollower represents a read-only follower of an XStore.
    properties:
      apiVersion:
        description: API version
        example: polardbx.aliyun.com/v1
        type: string
      kind:
        description: Resource kind
        example: XStoreFollower
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/xstores.ObjectMetaDTO'
        description: Resource metadata
      spec:
        allOf:
        - $ref: '#/definitions/xstores.XStoreFollowerSpecDTO'
        description: Desired state specification
      status:
        allOf:
        - $ref: '#/definitions/xstores.XStoreFollowerStatusDTO'
        description: Observed state status
    type: object
  xstores.XStoreFollowerSpecDTO:
    description: Desired state specification for XStore follower.
    properties:
      local:
        description: Local mode (no remote data center)
        example: true
        type: boolean
      resources:
        allOf:
        - $ref: '#/definitions/xstores.ResourceRequirementsDTO'
        description: Resource requirements
      role:
        description: Role of the follower node
        example: Learner
        type: string
      xstore:
        allOf:
        - $ref: '#/definitions/xstores.XStoreReferenceDTO'
        description: Reference to the source XStore
    type: object
  xstores.XStoreFollowerStatusDTO:
    description: Current observed state of XStore follower.
    properties:
      lagSeconds:
        description: Replication lag in seconds
        example: 0
        type: integer
      phase:
        description: Current phase
        example: Running
        type: string
      podName:
        description: Target pod name
        example: my-xstore-follower-0
        type: string
      ready:
        description: Ready status
        example: true
        type: boolean
    type: object
...
  xstores.XStoreReferenceDTO:
    description: Reference to the source XStore for backup.
    properties:
      name:
        description: Name of the XStore
        example: my-xstore
        type: string
      uid:
        description: UID of the XStore
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/xstores/{namespace}/{name}/rebuild/auto:
    post:
      description: Trigger automatic rebuild of XStore roles using internal heuristics.
      parameters:
      - description: Kubernetes namespace of the XStore
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the XStore
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Rebuild requested
          schema:
            $ref: '#/definitions/xstores.MessageResponseDTO'
        "404":
          description: XStore not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Auto rebuild XStore
      tags:
      - xstores
      - rebuild
  /api/v1/xstores/{namespace}/{name}/rebuild/cancel:
    delete:
      description: Cancel an ongoing rebuild operation for the specified XStore.
      parameters:
      - description: Kubernetes namespace of the XStore
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the XStore
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cancellation requested
          schema:
            $ref: '#/definitions/xstores.MessageResponseDTO'
        "404":
          description: Rebuild or XStore not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Cancel rebuild
      tags:
      - xstores
      - rebuild
  /api/v1/xstores/{namespace}/{name}/rebuild/learner:
    post:
      description: Trigger rebuild of the learner role for the specified XStore.
      parameters:
      - description: Kubernetes namespace of the XStore
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the XStore
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Rebuild requested
          schema:
            $ref: '#/definitions/xstores.MessageResponseDTO'
        "404":
          description: XStore not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Rebuild learner
      tags:
      - xstores
      - rebuild
  /api/v1/xstores/{namespace}/{name}/rebuild/logger:
    post:
      description: Trigger rebuild of the logger role for the specified XStore.
      parameters:
      - description: Kubernetes namespace of the XStore
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the XStore
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Rebuild requested
          schema:
            $ref: '#/definitions/xstores.MessageResponseDTO'
        "404":
          description: XStore not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Rebuild logger
      tags:
      - xstores
      - rebuild
  /api/v1/xstores/{namespace}/{name}/rebuild/progress:
    get:
      description: Get detailed progress information for an ongoing rebuild.
      parameters:
      - description: Kubernetes namespace of the XStore
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the XStore
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rebuild progress details
          schema:
            $ref: '#/definitions/xstores.RebuildProgressDTO'
        "404":
          description: Rebuild or XStore not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Get rebuild progress
      tags:
      - xstores
      - rebuild
  /api/v1/xstores/{namespace}/{name}/rebuild/status:
    get:
      description: Get status of an ongoing rebuild operation for the specified XStore.
      parameters:
      - description: Kubernetes namespace of the XStore
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the XStore
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rebuild status
          schema:
            $ref: '#/definitions/xstores.RebuildStatusDTO'
        "404":
          description: Rebuild or XStore not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Get rebuild status
      tags:
      - xstores
      - rebuild
  /api/v1/xstores/{namespace}/{name}/rebuild/wait:
    get:
      description: Block until a rebuild operation finishes or times out.
      parameters:
      - description: Kubernetes namespace of the XStore
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the XStore
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Final rebuild status
          schema:
            $ref: '#/definitions/xstores.RebuildStatusDTO'
        "404":
          description: Rebuild or XStore not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Wait rebuild
      tags:
      - xstores
      - rebuild
...
  /api/v1/xstores/followers:
    get:
      description: List follower XStores in the specified or default namespace.
      parameters:
      - description: Kubernetes namespace; defaults to 'default' when omitted
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of follower XStores
          schema:
            items:
              $ref: '#/definitions/xstores.XStoreFollowerDTO'
            type: array
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: List XStore followers
      tags:
      - xstores
      - followers
    post:
      consumes:
      - application/json
      description: Create a follower XStore that replicates from a primary XStore.
      parameters:
      - description: Kubernetes namespace; defaults to 'default' when omitted
        in: query
        name: namespace
        type: string
      - description: Follower XStore specification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/xstores.XStoreFollowerSpecDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created follower XStore
          schema:
            $ref: '#/definitions/xstores.XStoreFollowerDTO'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Create XStore follower
      tags:
      - xstores
      - followers
  /api/v1/xstores/followers/{namespace}/{name}:
    delete:
      description: Delete a follower XStore by namespace and name.
      parameters:
      - description: Kubernetes namespace of the follower
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the follower XStore
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: Deletion confirmation
          schema:
            $ref: '#/definitions/xstores.MessageResponseDTO'
        "404":
          description: Follower not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Delete XStore follower
      tags:
      - xstores
      - followers
    get:
      description: Get a follower XStore by namespace and name.
      parameters:
      - description: Kubernetes namespace of the follower
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the follower XStore
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Follower XStore
          schema:
            $ref: '#/definitions/xstores.XStoreFollowerDTO'
        "404":
          description: Follower not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Get XStore follower
      tags:
      - xstores
      - followers
    put:
      consumes:
      - application/json
      description: Update an existing follower XStore.
      parameters:
      - description: Kubernetes namespace of the follower
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the follower XStore
        in: path
        name: name
        required: true
        type: string
      - description: Updated follower specification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/xstores.XStoreFollowerSpecDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated follower XStore
          schema:
            $ref: '#/definitions/xstores.XStoreFollowerDTO'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Follower not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: Update XStore follower
      tags:
      - xstores
      - followers
swagger: "2.0"